requirements.txt
# TCFMamba dependencies (from conda env "mamba", 2026-02)
# Install: pip install -r requirements.txt
# PyTorch: install separately by platform (e.g. pip install torch --index-url https://download.pytorch.org/whl/cu121)

# Core
torch>=2.0.0
recbole>=1.2.0
einops>=0.7.0

# Data
numpy>=1.24.0,<2
pandas>=2.0.0
scipy>=1.10.0

# Config & utils
pyyaml>=6.0
tqdm>=4.66.0
tabulate>=0.9.0

# Logging & viz
tensorboard>=2.14.0
wandb>=0.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Mamba SSM (Linux/macOS from PyPI; Windows see WINDOWS_INSTALL.md)
causal-conv1d>=1.1.0
mamba-ssm>=1.1.0
